third_party/blink/web_tests/platform/android/external/wpt/html/semantics/forms/textfieldselection/selection-not-application-textarea-expected.txt

This is a testharness.js-based test.
FAIL text field selection for the input textarea assert_equals: initial selectionDirection expected "none" but got "forward"
Harness: the test ran to completion.
